Move kms-helper from lib to utils
